rename component to resource
