+++ client/app/bundles/shared/components/Resource/Resource.scss
